Changed the activity validator for fitness levels to use the int value of the Integer field, and changed the tests in user controller junit tests to account for the fitness level validations, an activity now has to include fitness levels before it is created and stored in the database #story[466]

+++ server/src/main/java/com/springvuegradle/seng302team600/validator/ActivityValidator.java
@@ -56,7 +56,7 @@ public class ActivityValidator {
      * @throws ResponseStatusException if invalid then bad request (400) should be returned
      */
     private static void validateFitness(Integer fitnessLevel) {
-        if (fitnessLevel > MAX_FITNESS_LEVEL || fitnessLevel < MIN_FITNESS_LEVEL) {
+        if (fitnessLevel.intValue() > MAX_FITNESS_LEVEL || fitnessLevel.intValue() < MIN_FITNESS_LEVEL) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Activity fitness level does not exist");
         }
     }
